refactor(E2eiEnrollment): move enrollment test utils into module
