+++ crypto/src/mls/conversation/mod.rs
@@ -598,7 +598,7 @@ mod tests {
     #[apply(all_cred_cipher)]
     #[wasm_bindgen_test]
     pub async fn create_many_people_conversation(case: TestCase) {
-        use crate::e2e_identity::test_utils::failsafe_ctx;
+        use crate::e2e_identity::enrollment::test_utils::failsafe_ctx;
 
         run_test_with_client_ids(case.clone(), ["alice"], move |[mut alice_central]| {
             Box::pin(async move {
